Adding support for alternate path templating engines.
This will allow people who prefer Rails-style, Selector-style, or any other style
URI templates to implement and use the library of their choice.  Only a minimal
interface is required for functionality (three methods and a constructor), and
nearly all of the in-built features can be supported (the one exception is that
resources cannot be nested beneath a route from a non-standard templating
engine; in practice, this issue should be rare).